Load the Code to CUI Map

The Code to CUI Map is provided as a dat file is written to the CODE_CUI_MAP table in the given schema.

load_code_cui_map(
  conn,
  schema = "nci",
  log_schema = "public",
  log_table = "setup_nci_log"
)

Arguments

conn
Postgres connection object.

schema
Schema to write to, Default: 'nci'

log_schema
Schema for the log table, Default: 'public'

log_table
Name of log table, Default: 'setup_nci_log'

Details

The dat file contains the map of NCI Thesaurus (NCIt) codes to NCI Metathesaurus (NCIMeta) CUIs (concept unique identifier). The CUIs are either derived from the UMLS Metathesaurus (CUIs that begin with "C-digit") or specific to the NCI Metathesaurus (CUIs that begin with "CL"). This mapping file is created for each release of the NCIMeta, which is published on a quarterly schedule hence is not current with every release of NCIt which is published monthly. Newly created Thesaurus concepts will not appear in this file until the next release of the Metathesaurus, and newly retired Thesaurus concepts will continue to appear until the next release of the Metathesaurus.

See also